requirements.txt
cozmo[camera]
pandas
numpy
opencv-python
matplotlib
Pillow
selenium
webdriver-manager